src/test/java/com/dexhacker/leetcode/remove_nodes_from_linked_list/SolutionTest.java
package com.dexhacker.leetcode.remove_nodes_from_linked_list;

import com.dexhacker.leetcode.structures.ListNode;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class SolutionTest {
    private Solution solution;

    @BeforeEach
    public void setUp() {
        solution = new Solution();
    }

    @Test
    public void testRemoveNodes() {
        ListNode node1 = new ListNode(5);
        ListNode node2 = new ListNode(2);
        ListNode node3 = new ListNode(13);
        ListNode node4 = new ListNode(3);
        ListNode node5 = new ListNode(8);
        node1.next = node2;
        node2.next = node3;
        node3.next = node4;
        node4.next = node5;

        ListNode result = solution.removeNodes(node1);
        assertEquals(result.val, 13);
        assertEquals(result.next.val, 8);
    }

    @Test
    public void testRemoveNodes2() {
        ListNode node1 = new ListNode(1);
        ListNode node2 = new ListNode(1);
        ListNode node3 = new ListNode(1);
        ListNode node4 = new ListNode(1);
        node1.next = node2;
        node2.next = node3;
        node3.next = node4;

        ListNode result = solution.removeNodes(node1);
        assertEquals(result.val, 1);
        assertEquals(result.next.val, 1);
        assertEquals(result.next.next.val, 1);
        assertEquals(result.next.next.next.val, 1);
    }
}
